pyproject.toml
[tool.poetry]
name = "transmission-rss"
version = "0.1.0"
description = ""
authors = ["Vlad Sudakov"]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.28.1"
feedparser = "^6.0.10"
transmission-rpc = "^3.4.0"
torrentool = "^1.1.1"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
black = "^22.12.0"
pytest = "^7.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
